Fixed spinner not showing correct week number
The spinner did not show the correct week number when scrolling between
different weeks. Also this is built using the new Android 5.0 SDK. A new
bug has been discovered - The positions of every schedule is offset one
week to the right in the viewpager, meaning that week 52 starts at
position 1. I will look into this bug as soon as I can