Fixed every known bug + transfered project to Android Studio
Every bug that I know of is fixed. The app is showing the correct day
and week at the correct position. The app is cleaned from useless code.
When swiping from last to first position the app no longer crashes. This
seems to be the first stable release of NovaScheduler!